AND NOW WE HAVE AN ACTUALLY ROOTLESS CONTAINER
Some considerations with regards to the pipe thing to communicate
between parent and child:

I considered both `channel` and `Arc<Mutex<()>>` to do that work, but it
turns out that neither of them plays nicely with the way arguments are
being sent using `clone`. Because I didn't want to debug much and this
is a simple use case, I just used pipes as the man pages for
`user_namespaces` examplifies.