Support detach of process
There's a little bit of noise here as I had to move the ownership of the
cgroups somewhere else. The basic problem is: If we detach from the
container, the `Cgroup` lifetime ends with the command it will try to
remove it. That will fail, but there would still be a process running in
that `Cgroup`. And worse, when that process finishes, the `Cgroup` won't
be deleted.

One might feel tempted to move the ownership inside the closure created
when we call `clone`, a problem with that is that after we `chroot` we
won't be able to remove it anymore.

A solution is to start a parent process whose sole purpose it's to
handle the lifetime of the cgroup. This takes a bit more memory, but
seems rather necessary.