Add three new commands.
- List running containers. It assumes that a container is running if and
only if its own cgroup it's still alive. I'm not sure how much of a
solid assumption this is, but I think it will work good enough.

- Delete a running container. Do so by killing all the processes under
the cgroiup of the container. We don't handle the errno `ESRCH` of this
sycall as an actual error because cgroups doesn't guarantee that all the
processes in this file would be running and because one of the
previously killed processes might have caused this one to be terminated.

- Output the logs of a container. It sends the output to `stdout` and
the error to `stderr`. TODO: Consider having separate commands for each
buffer. And rename `JailLogs` to `JailBuffer`, as it also includes
`stdin`, which isn't a log.