Create remaining parts of the import interface
1.- Building the layer stack. Basically, it grabs the
latest layer from the repositories file, puts it in a
stack, goes there and check its parent and if it exist, it
repeats the process.

2.- Creating the base subvolume: Simply call the
appropriate ioctl and then untar the layer in the
resulting path.

3,- Creating the next layer: A little bit more complex and
based on https://github.com/opencontainers/image-spec/blob/master/layer.md
It first creates a snapshot from the parent, then it
processes the entries in the tarball to find all the
whiteouts, the opaque whiteouts and the modification.
Finally, for every whiteout, it removes the file in the
in progress layer; for every opaque whiteout, it removes
all the files in the directory; and for every modification
it simply copies the file from the tarball into the
snapshot.

TODO: Don't create snapshot/subvolume if they already
exist, nor fail.