First steps on exporting
- Create and parse the new command to export

- Create a stack with the layers in the system. The top
element of the stack is the bottom layer. We get them by
using the `ioctl` syscall to get the information about the
subvolume and by contacting the father received there.
Once we receive a "Not found" error when contacting a
father, we have arrived to the parent of the tree (the
bottom layer).